Introduce libgsi, gsi_tool.
libgsi is a very small library to communicate critical gsi knowledge to
init and fs_mgr, to avoid hardcoding path names and constants.

gsi_tool is a shell wrapper to talk to gsid through binder.

gsid (currently in vold, will be moved shortly) is the daemon to manage
gsi installs.

Bug: 121210348
Test: N/A
Change-Id: I7942af56cb8e6ed69d28189bcf8d9806aa5e3e45