Add a progress bar API to gsid and use it in gsi_tool.
This patch introduces a getInstallProgress API call to IGsiService. It
is intended to be called while startGsiInstall or
commitGsiChunkFromStream are in progress, and will return details about
the current state. These details include the number of bytes left to
process and which step of the function is blocking.

Callers should invoke getInstallProgress on a separate thread. It is
safe to invoke when no operations are in progress. The caller must take
care to handle state transitions that are missed due to delays or sleep
calls. For example, call #1 may return step A at 50%, and call #2 may
return step B at 10%. It is the caller's responsibility to note that
the previous step successfully completed. gsi_tool contains sample code
for using this API.

Bug: 122556707
Test: manual test
Change-Id: I9af952368834bc3b5b5dadbbee6e91094d237cad