Force reboots within a GSI to return to the normal system image.
For now we are going to make GSI installs ephemeral. They can be
preserved (by not explicitly wiping them), but they will only get one
boot attempt total. If they fail to boot, or if the user reboots the
device once within the GSI, the next boot will return to the normal
system image. The existing GSI can then be re-enabled, or re-installed.

Bug: 122556707
Fixes: 123103015
Test: manual test
Change-Id: Iab07087fa5cd8b02af14bf0a5f2726018294f036