Start gsid on-demand.
With this patch, gsid is disabled by default. It can be enabled
on-demand via setting the ctl.start property to "gsid".

In addition, this patch cleans up error reporting in gsi_tool to
distinguish between binder and IGsiService errors, and fixes a bug where
"gsi_tool enable" would not work from a non-root shell.

Bug: 126622385
Test: gsid not running by default
Change-Id: Icd6350f146d12e71befafa91ac859cf438376bc1