Track image sizes separately from image files.
If FiemapWriter needs to bump up the image file size due to alignment,
gsid must track the actual requested size separately for AVB to work.
Otherwise, libavb will be unable to find the vbmeta footer at the end of
the partition.

The actual size can be recovered from /metadata/gsi/lp_metadata, as long
as the extent list is truncated appropriately.

Bug: 126230649
Test: gsi install works
Change-Id: If5b580691d9399dfefceef89cc7e574d422f06e2