Do not hardcode /data/gsi in GsiService.
This patch removes hardcoding of /data/gsi. Instead, StartInstall is
given an installation directory, and this is saved in
/metadata/gsi/install_dir. When wiping or re-enabling a GSI, image paths
are recovered from install_dir.

This does not actually make the path configurable: it is a refactoring
to no longer rely on hardcoded image paths.

Included are some minor cleanups to ensure that cancelled installs
delete files properly.

Bug: 126230649
Test: gsi_tool install, enable, wipe
Change-Id: Ieb788e84d19127e8e3375909f3c522e16462d384