Don't call HasPinnedExtents when wrapping with device-mapper.
With the recent changes to f2fs and libfiemap_writer, HasPinnedExtents
now returns false if a file was never pinned (even if it hasn't been
moved). This exposed a bug in IGsiService. When using device-mapper to
write the GSI, we checked HasPinnedExtents() on the underlying file
path. However this path was to the split image header (i.e.
system_gsi.img), rather than the contents (system_gsi.img.0001).

The simplest fix is to just remove this check. A more accurate check is
performed in SetGsiBootable.

Bug: 129803146
Test: gsi_tool install works
Change-Id: I9cf996ebc054147612df696dc73cc6d8df04f44d
Merged-In: I9cf996ebc054147612df696dc73cc6d8df04f44d