Fix getGsiBootStatus according to framework's usage
After install AoT and boot twice. The device will be in a
installed/disabled state. In this state, if the user uses the
notification bar to [RESTART] AoT. The getGsiBootStatus will return a
"Disabled" state because AoT is not running then the device will be
trapped in AoT mode until the user uses gsi_tool to disable it
explicitly.

Test: gsi_tool install & reboot twice & re-enable from notification bar
Bug: 131120650
Change-Id: I5d1bd607ad5e811348fc8caa8359abb7b51b0bb7