Support ReenableGsi when IsGsiRunning is true
This is required by the AoT Compliance test. The current design
provides the capability to re-enable gsi but with an implicit
limitation that this method can not be called in the running state.
It makes our test engineers need to reboot device twice in order to
pin the device in AoT mode, e.g. go back to the original system and
gsi_tool enable without the -s option. This CL can decrease the
setup time by removing this limitation.

Test: "gsi_tool enable" when gsi_tool status shows running
Bug: 131383526
Change-Id: Iecf4c1dddb07b1261117011813022966660f10bc