libfiemap: Simplify block device derivation in MapImageDevice.
During testing, we can create fiemap images on non-data block devices,
such as /dev/block/loopN. When this happens, the "super" name in
lp_metadata is "loopN", which does not work with PartitionOpener. We
worked around that by instead probing the actual block device and then
passing it to MapWithDmLinear.

This is pretty edge-casey, so instead, use the normal flow for
CreateLogicalPartition and don't override the block device. To make this
work during testing, we allow overriding the PartitionOpener in
ImageManager. The test can then prepend /dev/block as needed.

Bug: 139204329
Test: fiemap_image_test gtest
Change-Id: I88d50e0668bc535ea2632a8b01a7d2eeb29e9816