Enhance the fault tolerance of GSID installation
The gsid may function abnormally when an installation get interrupted.

case 1.
$ gsi_tool install

 ... use ctrl+C to stop ...

$ gsi_tool wipe
Could not remove GSI install.

case 2
When it's interrupted. The installer destructor does not have the
the chance to run therefore it can be a resource leak because the
backing files, e.g. system_gsi and userdata_gsi are not released.
Neither can they be enabled because the IsGsiInstalled() == false.

This CL fixes cases above by making a integrity model based on
IsGsiInstalled() as an atomic switch, e.g.

if (IsGsiInstalled()) {
    The installation must be complete and the enableGsi() must work.
} else {
    All resources that have been allocated must be cleaned up.
    a. Clean up methods and command must work.
    b. runStartup method must do the clean up.
}

All corresponding changes are based on this model.

Bug: 140014207
Test: \
 1. gsi_tool install & reboot
 2. gsi_tool install & ctrl+C & gsi_tool status & gsi_tool wipe & gsi_tool status
 3. gsi_tool install & ctrl+C & gsi_tool status & reboot & gsi_tool status

Change-Id: I0dcf043d65b50f0dd82eac1f428a295e94c81812