libfiemap: Allow a larger number of FIEMAP extents.
The 512 extent limit is too low and it's easy to go past it, especially
when rapidly deleting and installing DSUs. First, bump the limit to
something much higher (50000 picked here as we've tested dm-linear with
this). Next, refactor ReadFiemap so it first queries the exact number of
extents, rather than passing a fixed-size buffer.

Bug: 144424457
Test: fiemap_writer_test gtest
Change-Id: I9da44b4e199d178d184ba53f526d8b4c518f0675