Only allow certain CreateBackingImage requests when pinning is fully supported.
F2FS does not yet support reliable pinning without patching. We support
this for DSUs (and testing), but for libsnapshot and adb remount we need
a guarantee that blocks will not be moved.

Note that the check here is a sanity check, and pin_file support will be
further checked in VTS.

Bug: 134949511
Test: manual test
Change-Id: Ida54964eeecbd50d75d157c4aa89d87225d15ac8