Add "create-partition" debugging API to gsi_tool
`gsi_tool create-partition [-p <name>] [-s <size>] [--readwrite]`
is a debugging API used to create DSU partitions. Readwritable
partitions such as userdata and scratch should be created with the
"--readwrite" flag.

This is like `gsi_tool install` but `create-partition` is more primitive
and does less things (doesn't auto-create userdata, doesn't auto-reboot.)

This is a debugging API so don't depend on its behavior and don't expect
it to be stable.

Bug: 165925766
Test: adb shell gsi_tool create-partition --readwrite \
  --partition scratch --size $((200 * 1024 * 1024))
Change-Id: I9ce463220534ba46c948ea48e9c52bdc0ff762b1