Create the backing file for the PersistentDataBlock
In order to support the test harness mode in DSU. A copy-on-write
mechanism is added to the system server to protect the on-device
PersistentDataBlock from any modification. When running a DSU, all
write operations are redirected to a backing file. This CL creates
the backing file with rc file to get the correct se file label.

Bug: 175852148
Test: gsi_tool install

Change-Id: I739e28a4cbc36330e1ba39de0e8b7d5a111617f7