Revert "GsiService::enableGsi() checks return code of FinishInstall()"
This reverts commit 99ab518ed6b7fae95ba29b132227f993e0a099d0.

enableGsi() shouldn't check FinishInstall() explicitly, instead the
caller of enableGsi() should call closePartition(), and enableGsi()
should assume that the installation was successful and any error was
handled by the caller already.
closePartition() was added in aosp/1404709.
This opens up the path for optional DSU partitions. For example, failing
to install a overlayFS scratch partition shouldn't be fatal, and the DSU
installation should be able to continue without a scratch.

Bug: 165925766
Bug: 165471299
Test: TH
Change-Id: Ic9b653426cf7fc3c92279e9ceb7c1eb662c15cef