closeInstall() and enableGsi() code cleanup
Right now the following workflow doesn't work,
$ adb shell am start-activity \
  -n com.android.dynsystem/com.android.dynsystem.VerificationActivity \
  -a android.os.image.action.START_INSTALL \
  -d file:///sdcard/system.img.zip \
  --el KEY_USERDATA_SIZE $(( 8 << 30 ))
(wait until completion)
$ adb reboot
$ adb shell am start-service \
  -n com.android.dynsystem/com.android.dynsystem.DynamicSystemInstallationService \
  -a com.android.dynsystem.ACTION_REBOOT_TO_DYN_SYSTEM

because the DSU installation would be discarded after the reboot.
This is because some of the installation complete marker files are not
created by closeInstall(), but confusingly by enableGsi().

This change straightens the code of closeInstall() and enableGsi() and
make closeInstall() do actually what it claims to do, that is to "mark
an installation completion" and enableGsi() should only do one job, that
is to "enabled a completed installation".

Bug: 165471299
Test: Install DSU, adb reboot, and then reboot to DSU
Test: Install DSU and immedeitaly reboot to DSU
Change-Id: Ieeff9617df97e1a5348b4f9c7918427e825954c6