DsuGsiIntegrationTest: Verify that storage space are freed after 'wipe'
'gsi_tool wipe' should free up any space used by DSU.
We test this by comparing the free space of /data before and after
'wipe'. As there could be other processes writing to the /data partition
at the same time this test is running, the test should be fuzzy to
account for the dynamic nature of /data. Use 0.9 (allow 10% jitter) as
the dampening factor for the test. The number is chosen randomly.

Bug: 240389742
Test: DsuGsiIntegrationTest
Change-Id: I43580a50e4e897819157b5ec229d6997212aed6f