DsuGsiIntegrationTest: Add option to override the DSU userdata size
In order to run this test on virtual test bed, which has much smaller
storage space than physical devices, we need to run virtual test
configuration with smaller userdata size. This is not ideal, as
installing a DSU installation with tiny userdata is not realistic, but
it's a compromise we had to make in order to run this test continuously.

Bug: 240389742
Test: DsuGsiIntegrationTest DSUEndtoEndTest
Change-Id: I7cb898c00e7380bd2867d7995742f9f31d3563e6