gsi_tool: gsid->openImageService() could fail if already running in DSU
The problematic line is
https://cs.android.com/android/platform/superproject/+/master:system/gsid/gsi_service.cpp;l=807;drc=352caad81dcd4a39bf9c0dd53661af2e75803b30

After aosp/2577792 is merged, the DSU installation path is changed, thus
the above line would fail if DSU is already running. Because the
"data_dir" is referring to a path in the original "outside" /data. When
running inside DSU, the "data_dir" would be empty or doesn't exist.

Teach `gsi_tool status` to handle openImageService() failures in this
case.

Bug: 280831236
Test: abtd dsu_gsi_integration_test
Change-Id: Iaf6909d3ab4ec0b3d399437fdf3644b3e7ad8e5e